auto-report-generator/requirements.txt
streamlit
pandas
openai
PyPDF2
python-docx
matplotlib
plotly
pdfkit
xlsxwriter
numpy
requests
